Move Present from package org.scalactic.anyvals to org.scalactic, and added a pretty toString.
